Fixed persistence. Fixed Add/remove buddy option

+++ src/main/java/org/example/AddressBookRepository.java
@@ -1,10 +1,8 @@
 package org.example;
-import java.util.List;
 
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressBookRepository extends JpaRepository<AddressBook, Long> {
+public interface AddressBookRepository extends JpaRepository<AddressBook, Integer> {
     AddressBook findById(long id);
     AddressBook findAddressBooksById(int id);
 }
